don't print "info depth x" lines without further info
all other engines only print the line together with the pv and other
info. since not all engines support "go depth", my tournament program
sends "stop" as soon as it sees a line with depth > desired depth,
so sungorus was at a disadvantage of 1 depth level due to this
superfluous information.